requirements.txt
pygame-ce>=2.4.0
numpy>=1.26.3
mypy>=1.8.0 # For type checking
black>=23.12.0 # For code formatting
pytest>=7.4.0 # For unit testing 